@@ -3,46 +3,49 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "build": "next build",
-    "dev": "next dev",
-    "lint": "next lint",
-    "start": "next start"
+    "build": "next build --webpack",
+    "dev": "next dev --webpack",
+    "lint": "eslint",
+    "start": "next start --webpack"
   },
   "dependencies": {
-    "axios": "^1.13.2",
+    "axios": "^1.13.5",
     "chart.js": "^4.5.1",
     "clsx": "^2.1.1",
     "luxon": "^3.7.2",
-    "mongodb": "^6.21.0",
-    "next": "^15.5.9",
+    "mongodb": "^7.1.0",
+    "next": "^16.1.6",
     "normalize.css": "^8.0.1",
-    "ramda": "^0.30.1",
-    "react": "^19.2.3",
+    "ramda": "^0.32.0",
+    "react": "^19.2.4",
     "react-chartjs-2": "^5.3.1",
-    "react-dom": "^19.2.3",
+    "react-dom": "^19.2.4",
     "react-icons": "^5.5.0",
     "sharp": "^0.34.5",
-    "uuid": "^11.1.0",
     "yup": "^1.7.1"
   },
   "devDependencies": {
-    "@next/bundle-analyzer": "^15.5.9",
+    "@eslint/js": "^10.0.1",
+    "@next/bundle-analyzer": "^16.1.6",
+    "@next/eslint-plugin-next": "^16.1.6",
     "@types/luxon": "^3.7.1",
-    "@types/node": "^22.19.2",
-    "@types/ramda": "^0.30.2",
-    "@types/react": "^19.2.7",
+    "@types/node": "^24.10.13",
+    "@types/ramda": "^0.31.1",
+    "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
-    "@types/uuid": "^10.0.0",
-    "@typescript-eslint/eslint-plugin": "^8.49.0",
-    "@vanilla-extract/css": "^1.17.5",
-    "@vanilla-extract/next-plugin": "^2.4.16",
-    "eslint": "^9.39.1",
-    "eslint-config-next": "^15.5.9",
+    "@typescript-eslint/eslint-plugin": "^8.56.1",
+    "@vanilla-extract/css": "^1.18.0",
+    "@vanilla-extract/next-plugin": "^2.4.17",
+    "eslint": "^9.39.3",
+    "eslint-config-next": "^16.1.6",
     "eslint-config-prettier": "^10.1.8",
-    "eslint-plugin-prettier": "^5.5.4",
-    "prettier": "^3.7.4",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-prettier": "^5.5.5",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^7.0.1",
+    "prettier": "^3.8.1",
     "typescript": "^5.9.3",
-    "typescript-eslint": "^8.49.0"
+    "typescript-eslint": "^8.56.1"
   },
-  "packageManager": "pnpm@10.8.1+sha512.c50088ba998c67b8ca8c99df8a5e02fd2ae2e2b29aaf238feaa9e124248d3f48f9fb6db2424949ff901cffbb5e0f0cc1ad6aedb602cd29450751d11c35023677"
+  "packageManager": "pnpm@10.30.3"
 }
