@@ -5,26 +5,32 @@ import { Metadata } from "next";
 import { notFound } from "next/navigation";
 import { cache } from "react";
 
+const getCertificate = cache(
+  async (id: string): Promise<CertificationType | null> => {
+    try {
+      const ret = await axios.get<CertificationType>(
+        `${process.env.NEXT_PUBLIC_API}/api/certificate/${id}`
+      );
+
+      return ret.data;
+    } catch {
+      return null;
+    }
+  }
+);
+
 export async function generateMetadata({ params }: Props): Promise<Metadata> {
   const { id } = await params;
 
-  try {
-    const cert = await getCertificate(id);
-    return {
-      title: `Certificate ${cert.id} - Certification Program - Dr. Kobros`
-    };
-  } catch {
+  const cert = await getCertificate(id);
+  if (!cert) {
     notFound();
   }
+
+  return {
+    title: `Certificate ${cert.id} - Certification Program - Dr. Kobros`
+  };
 }
-
-const getCertificate = cache(async (id: string): Promise<CertificationType> => {
-  const ret = await axios.get<CertificationType>(
-    `${process.env.NEXT_PUBLIC_API}/api/certificate/${id}`
-  );
-
-  return ret.data;
-});
 
 type Props = {
   params: Promise<{
@@ -35,14 +41,14 @@ type Props = {
 export default async function CertificationViewPage({ params }: Props) {
   const { id } = await params;
 
-  try {
-    const cert = await getCertificate(id);
-    return (
-      <>
-        <Certification certificate={cert} />
-      </>
-    );
-  } catch {
+  const cert = await getCertificate(id);
+  if (!cert) {
     notFound();
   }
+
+  return (
+    <>
+      <Certification certificate={cert} />
+    </>
+  );
 }
